requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
anthropic==0.39.0
pillow==10.2.0
python-dotenv==1.0.0
backoff==2.2.1
httpx
anyio
distro
jiter
pydantic
sniffio
typing-extensions
cloudinary==1.41.0
certifi>=2024.2.2
six
urllib3
requests==2.31.0
aiohttp==3.9.1
mysqlclient==2.2.1
psutil==5.9.8